requirements.txt
beautifulsoup4==4.12.2
deepl==1.14.0
playwright==1.34.0
python-dotenv==1.0.0
requests==2.31.0
typer==0.9.0
